dont test signal on macos because pkill does not work there

+++ util.c
@@ -1523,18 +1523,20 @@ main(int argc, char **argv) {
     char *bases[] = {
         "cccc", "cc", "c", "c", "cccc", "cccc", "cccc",
     };
-    struct sigaction signal_action;
-
     (void)argc;
-    signal_action.sa_handler = signal_handler;
-    sigemptyset(&signal_action.sa_mask);
-    signal_action.sa_flags = SA_RESTART;
-    if (sigaction(SIGUSR1, &signal_action, NULL) != 0) {
-        error2("Error in sigaction: %s.\n", strerror(errno));
-        exit(EXIT_FAILURE);
+
+    if (OS_LINUX) {
+        struct sigaction signal_action;
+        signal_action.sa_handler = signal_handler;
+        sigemptyset(&signal_action.sa_mask);
+        signal_action.sa_flags = SA_RESTART;
+        if (sigaction(SIGUSR1, &signal_action, NULL) != 0) {
+            error2("Error in sigaction: %s.\n", strerror(errno));
+            exit(EXIT_FAILURE);
+        }
+        send_signal(argv[0], SIGUSR1);
+        ASSERT(received_signal);
     }
-    send_signal(argv[0], SIGUSR1);
-    ASSERT(received_signal);
 
     memset64(p1, 0, SIZEMB(1));
     memcpy64(p1, string, strlen64(string));
